requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
discord.py==1.7.3
dnspython==2.1.0
idna==3.2
multidict==5.1.0
pdf2image==1.16.0
Pillow==8.3.0
pymongo==3.11.4
PyMuPDF==1.18.14
requests==2.25.1
typing-extensions==3.10.0.0
urllib3==1.26.6
yarl==1.6.3
